+++ frontend/src/app/components/BurnerWallet/BurnerWallet.tsx
@@ -1,12 +1,12 @@
 "use client";
 import { useEffect, useState } from "react";
-import AssetTransferModal from "../AssetTransferModal";
-import ConnectionModal from "../ConnectionModal";
+import AssetTransferModal from "./AssetTransferModal";
+import ConnectionModal from "./ConnectionModal";
 import { useContractRead } from "@starknet-react/core";
 import { Account, RpcProvider } from "starknet";
 import CopyButton from "../ui_components/util/CopyButton";
 import Erc20Abi from "../../abi/token.abi.json";
-import ContractExecutionModal from "../ContractExecutionModal";
+import ContractExecutionModal from "./ContractExecutionModal";
 import { ETH_SEPOLIA, STRK_SEPOLIA } from "@/app/utils/constant";
 import { formatCurrency } from "@/app/utils/currency";
 import AccountBalance from "../ui_components/AccountBalance";
